Refatorando testes, conforme solicitacao do instrutor. Incluido os testes de funcionalidade do CastMember.
