+++ app/Http/Controllers/Api/GenreController.php
@@ -22,7 +22,6 @@ class GenreController extends BasicCrudController
     protected function rulesUpdate()
     {
         return [
-            'id' => 'required',
             'name' => 'required|max:255',
             'is_active' => 'boolean'
         ];
