+++ routes/api.php
@@ -18,4 +18,5 @@
 Route::group(['namespace' => 'Api'], function () {
     Route::resource('categories', 'CategoryController', ['except' => ['create', 'edit']]);
     Route::resource('genres', 'GenreController', ['except' => ['create', 'edit']]);
+    Route::resource('cast_members', 'CastMemberController', ['except' => ['create', 'edit']]);
 });
