+++ tests/Feature/Http/Controllers/Api/GenreControllerTest.php
@@ -2,118 +2,155 @@
 
 namespace Tests\Feature\Http\Controllers\Api;
 
-use App\Http\Controllers\Api\BasicCrudController;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
-use Illuminate\Http\Request;
-use Illuminate\Validation\ValidationException;
-use ReflectionClass;
-use Tests\Stubs\Controllers\GenreControllerStub;
-use Tests\Stubs\Models\GenreStub;
+use App\Models\Genre;
+use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Tests\TestCase;
+use Tests\Traits\TestValidations;
 
 class GenreControllerTest extends TestCase
 {
-    /** @var GenreControllerStub $controller */
-    private $controller;
-    /** @var GenreStub $controller */
+    use DatabaseMigrations, TestValidations;
+
     private $genre;
 
-    protected function setUp(): void
+    public function setUp(): void
     {
         parent::setUp();
-        GenreStub::dropTable();
-        GenreStub::createTable();
-        $this->controller = new GenreControllerStub();
-        $this->genre = GenreStub::create(['name' => 'test_name']);
-    }
-
-    protected function tearDown(): void
-    {
-        GenreStub::dropTable();
-        parent::tearDown();
+        $this->genre = factory(Genre::class)->create();
+        $this->genre->refresh();
     }
 
     public function testIndex()
     {
-        $result = $this->controller->index()->toArray();
-        $this->assertEquals([$this->genre->toArray()], $result);
-    }
+        $response = $this->get(route('genres.index'));
 
-    public function testInvalidationDataInStore()
-    {
-        $this->expectException(ValidationException::class);
-        /** @var Request $request */
-        $request = \Mockery::mock(Request::class);
-        $request->shouldReceive('all')
-            ->once()
-            ->andReturn(['name' => '']);
-        $this->controller->store($request);
-    }
-
-    public function testStore()
-    {
-        /** @var Request $request */
-        $request = \Mockery::mock(Request::class);
-        $request->shouldReceive('all')
-            ->once()
-            ->andReturn(['name' => 'test_store_function']);
-        $obj = $this->controller->store($request);
-
-        $this->assertEquals(
-            GenreStub::where(['name' => 'test_store_function'])->firstOrFail()->toArray(),
-            $obj->toArray()
-        );
-    }
-
-    public function testIfFindOrFailFetchModel()
-    {
-        $reflectionClass = new ReflectionClass(BasicCrudController::class);
-        $reflectionMethod = $reflectionClass->getMethod('findOrFail');
-        $reflectionMethod->setAccessible(true);
-
-        $result = $reflectionMethod->invokeArgs($this->controller, [$this->genre->id]);
-        $this->assertInstanceOf(GenreStub::class, $result);
-    }
-
-    public function testIfFindOrThrowExceptionWhenIdInvalid()
-    {
-        $this->expectException(ModelNotFoundException::class);
-        $reflectionClass = new ReflectionClass(BasicCrudController::class);
-        $reflectionMethod = $reflectionClass->getMethod('findOrFail');
-        $reflectionMethod->setAccessible(true);
-
-        $result = $reflectionMethod->invokeArgs($this->controller, [0]);
-        $this->assertInstanceOf(GenreStub::class, $result);
+        $response
+            ->assertStatus(200)
+            ->assertJson([$this->genre->toArray()]);
     }
 
     public function testShow()
     {
-        $response = $this->controller->show($this->genre->id);
+        $response = $this->get(route('genres.show', ['genre' => $this->genre->id]));
 
-        $this->assertEquals(
-            GenreStub::find(1)->toArray(),
-            $response->toArray()
-        );
+        $response
+            ->assertStatus(200)
+            ->assertJson($this->genre->toArray());
+    }
+
+    public function testValidationData()
+    {
+        $this->validationStore();
+        $this->validationUpdate();
+    }
+
+    private function validationStore()
+    {
+        $data = ['name' => ''];
+        $this->assertInvalidationFieldsInStoreAction($data, 'required');
+
+        $data = ['name' => str_repeat('a', 256)];
+        $this->assertInvalidationFieldsInStoreAction($data, 'max.string', ['max' => 255]);
+
+        $data = ['is_active' => 'a'];
+        $this->assertInvalidationFieldsInStoreAction($data, 'boolean');
+    }
+
+    private function validationUpdate()
+    {
+        $data = ['name' => ''];
+        $this->assertInvalidationFieldsInUpdateAction($data, 'required');
+
+        $data = ['name' => str_repeat('a', 256)];
+        $this->assertInvalidationFieldsInUpdateAction($data, 'max.string', ['max' => 255]);
+
+        $data = ['is_active' => 'a'];
+        $this->assertInvalidationFieldsInUpdateAction($data, 'boolean');
+    }
+
+    public function testStore()
+    {
+        $response = $this->json('POST', route('genres.store'), [
+            'name' => 'test'
+        ]);
+        $id = $response->json('id');
+        $genre = Genre::find($id);
+
+        $response
+            ->assertStatus(201)
+            ->assertJson($genre->toArray());
+        $this->assertTrue($response->json('is_active'));
+
+        $response = $this->json('POST', route('genres.store'), [
+            'name' => 'test'
+        ]);
+        $id = $response->json('id');
+        $genre = Genre::find($id);
+
+        $response
+            ->assertJsonFragment(['is_active' => true]);
     }
 
     public function testUpdate()
     {
-        $request = \Mockery::mock(Request::class);
-        $request->shouldReceive('all')
-            ->once()
-            ->andReturn(['name' => 'test_name_changed']);
-        $obj = $this->controller->update($request, $this->genre->id);
-        $this->assertEquals(
-            GenreStub::find(1)->toArray(),
-            $obj->toArray()
+        $response = $this->json(
+            'PUT',
+            route(
+                'genres.update',
+                ['genre' => $this->genre->id]
+            ),
+            [
+                'name' => 'name_altered',
+                'is_active' => true
+            ]
         );
+
+        $genre = Genre::find($response->json('id'));
+        $response
+            ->assertStatus(200)
+            ->assertJson($genre->toArray())
+            ->assertJsonFragment([
+                'name' => 'name_altered',
+                'is_active' => true
+            ]);
+
+
+        $genre = factory(Genre::class)->create([
+            'is_active' => false
+        ]);
+        $response = $this->json(
+            'PUT',
+            route(
+                'genres.update',
+                ['genre' => $genre->id]
+            ),
+            [
+                'name' => 'name_altered',
+                'is_active' => true
+            ]
+        );
+
+        $genre = Genre::find($response->json('id'));
     }
 
     public function testDestroy()
     {
-        $response = $this->controller->destroy($this->genre->id);
-        $this->createTestResponse($response)
-            ->assertStatus(204);
-        $this->assertCount(0, GenreStub::all());
+        $response = $this->json(
+            'DELETE',
+            route('genres.destroy', ['genre' => $this->genre->id])
+        );
+        $response->assertNoContent();
+        $this->assertNull(Genre::find($this->genre->id));
+        $this->assertNotNull(Genre::withTrashed()->find($this->genre->id));
+    }
+
+    protected function routeStore()
+    {
+        return route('genres.store');
+    }
+
+    protected function routeUpdate()
+    {
+        return route('genres.update', ['genre' => $this->genre->id]);
     }
 }
